+++ tests/it/main.rs
@@ -3,6 +3,7 @@ mod cancel;
 mod common;
 mod error;
 mod gat_resume;
+mod invoke;
 mod macros;
 mod non_static;
 mod proc_macros;